Add ApiController and corresponding template
Added a new ApiController with a main route to src/Controller/api/v1, along with a corresponding Twig template. These changes initiate the basic structure for the v1 of our API subsequently providing a foundation for building out the API's endpoints. The template contains necessary pointers to controller and template file paths easing the debugging process.